# docs/swagger.yaml
definitions:
  dao.AccessControl:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      id:
        type: integer
      open_auth:
        type: integer
      service_flow_limit:
        type: integer
      service_id:
        type: integer
      white_host_name:
        type: string
      white_list:
        type: string
    type: object
  dao.App:
    properties:
      app_id:
        type: string
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      update_at:
        type: string
      white_ips:
        type: string
    type: object
  dao.GrpcRule:
    properties:
      header_transfor:
        type: string
      id:
        type: integer
      port:
        type: integer
      service_id:
        type: integer
    type: object
  dao.HttpRule:
    properties:
      header_transfor:
        type: string
      id:
        type: integer
      need_https:
        type: integer
      need_strip_uri:
        type: integer
      need_websocket:
        type: integer
      rule:
        type: string
      rule_type:
        type: integer
      service_id:
        type: integer
      url_rewrite:
        type: string
    type: object
  dao.LoadBalance:
    properties:
      check_interval:
        type: integer
      check_method:
        type: integer
      check_timeout:
        type: integer
      forbid_list:
        type: string
      id:
        type: integer
      ip_list:
        type: string
      round_type:
        type: integer
      service_id:
        type: integer
      upstream_connect_timeout:
        type: integer
      upstream_header_timeout:
        type: integer
      upstream_idle_timeout:
        type: integer
      upstream_max_idle:
        type: integer
      weight_list:
        type: string
    type: object
  dao.ServiceDetail:
    properties:
      access_control:
        $ref: '#/definitions/dao.AccessControl'
      grpc_rule:
        $ref: '#/definitions/dao.GrpcRule'
      http_rule:
        $ref: '#/definitions/dao.HttpRule'
      info:
        $ref: '#/definitions/dao.ServiceInfo'
      load_balance:
        $ref: '#/definitions/dao.LoadBalance'
      tcp_rule:
        $ref: '#/definitions/dao.TcpRule'
    type: object
  dao.ServiceInfo:
    properties:
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      load_type:
        type: integer
      service_desc:
        type: string
      service_name:
        type: string
      update_at:
        type: string
    type: object
  dao.TcpRule:
    properties:
      id:
        type: integer
      port:
        type: integer
      service_id:
        type: integer
    type: object
  dto.AdminInfoOutput:
    properties:
      avatar:
        type: string
      id:
        type: integer
      introduction:
        type: string
      login_time:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  dto.AdminLoginInput:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.AdminLoginOutput:
    properties:
      token:
        description: token
        example: token
        type: string
    type: object
  dto.AppAddInput:
    properties:
      app_id:
        type: string
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      white_ips:
        type: string
    required:
    - app_id
    - name
    type: object
  dto.AppListItemOutput:
    properties:
      app_id:
        type: string
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      real_qpd:
        type: integer
      real_qps:
        type: integer
      secret:
        type: string
      update_at:
        type: string
      white_ips:
        type: string
    type: object
  dto.AppListOutput:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.AppListItemOutput'
        type: array
      total:
        type: integer
    type: object
  dto.AppUpdateInput:
    properties:
      app_id:
        type: string
      id:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      white_ips:
        type: string
    required:
    - app_id
    - id
    - name
    type: object
  dto.ChangePwdInput:
    properties:
      password:
        example: "123456"
        type: string
    required:
    - password
    type: object
  dto.DashServiceStatItemOutput:
    properties:
      load_type:
        type: integer
      name:
        type: string
      value:
        type: integer
    type: object
  dto.DashServiceStatOutput:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DashServiceStatItemOutput'
        type: array
      legend:
        items:
          type: string
        type: array
    type: object
  dto.PanelGroupDataOutput:
    properties:
      appNum:
        type: integer
      currentQps:
        type: integer
      serviceNum:
        type: integer
      todayRequestNum:
        type: integer
    type: object
  dto.ServiceAddGRPCInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      header_transfor:
        type: string
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        maximum: 8999
        minimum: 8001
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceAddHTTPInput:
    type: object
  dto.ServiceAddTCPInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      header_transfor:
        type: string
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        maximum: 8999
        minimum: 8001
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceInfoItem:
    properties:
      id:
        example: 1
        type: integer
      load_type:
        example: 1
        type: integer
      qpd:
        type: integer
      qps:
        type: integer
      service_addr:
        type: string
      service_desc:
        example: ""
        type: string
      service_name:
        example: test_grpc
        type: string
      total_node:
        type: integer
    type: object
  dto.ServiceInfoOutput:
    properties:
      info_list:
        items:
          $ref: '#/definitions/dto.ServiceInfoItem'
        type: array
      total:
        type: integer
    type: object
  dto.ServiceUpdateGRPCInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      header_transfor:
        type: string
      id:
        type: integer
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        maximum: 8999
        minimum: 8001
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - id
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceUpdateHTTPInput:
    type: object
  dto.ServiceUpdateTCPInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      id:
        type: integer
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        maximum: 8999
        minimum: 8001
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - id
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
  middleware.Response:
    properties:
      data: {}
      errmsg:
        type: string
      errno:
        $ref: '#/definitions/middleware.ResponseCode'
      stack: {}
      trace_id: {}
    type: object
  middleware.ResponseCode:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 401
    - 1000
    - 2001
    type: integer
    x-enum-varnames:
    - SuccessCode
    - UndefErrorCode
    - ValidErrorCode
    - InternalErrorCode
    - InvalidRequestErrorCode
    - CustomizeCode
    - GROUPALL_SAVE_FLOWERROR
info:
  contact: {}
paths:
  /admin/admin_info:
    get:
      consumes:
      - application/json
      description: 管理员信息
      operationId: /admin/admin_info
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminInfoOutput'
              type: object
      summary: 管理员信息
      tags:
      - 管理员接口
  /admin/change_pwd:
    post:
      consumes:
      - application/json
      description: 密码修改
      operationId: /admin/change_pwd
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePwdInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 密码修改
      tags:
      - 管理员接口
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: 管理员登陆
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminLoginOutput'
              type: object
      summary: 管理员登陆
      tags:
      - 管理员接口
  /admin_login/logout:
    get:
      consumes:
      - application/json
      description: 管理员退出
      operationId: /admin_login/logout
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 管理员退出
      tags:
      - 管理员接口
  /app/app_delete:
    get:
      consumes:
      - application/json
      description: 租户删除
      operationId: /app/app_delete
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 租户删除
      tags:
      - 租户管理
  /app/app_detail:
    get:
      consumes:
      - application/json
      description: 租户详情
      operationId: /app/app_detail
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dao.App'
              type: object
      summary: 租户详情
      tags:
      - 租户管理
  /app/app_info_add:
    post:
      consumes:
      - application/json
      description: 租户添加
      operationId: /app/app_info_add
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AppAddInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 租户添加
      tags:
      - 租户管理
  /app/app_list:
    get:
      consumes:
      - application/json
      description: 租户列表
      operationId: /app/app_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页多少条
        in: query
        name: page_size
        required: true
        type: integer
      - description: 页码
        in: query
        name: page_no
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AppListOutput'
              type: object
      summary: 租户列表
      tags:
      - 租户管理
  /app/app_update:
    post:
      consumes:
      - application/json
      description: 租户更新
      operationId: /app/app_update
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AppUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 租户更新
      tags:
      - 租户管理
  /dashboard/panel_group_data:
    get:
      consumes:
      - application/json
      description: 指标统计
      operationId: /dashboard/panel_group_data
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PanelGroupDataOutput'
              type: object
      summary: 指标统计
      tags:
      - 首页大盘
  /dashboard/service_stat:
    get:
      consumes:
      - application/json
      description: 服务统计
      operationId: /dashboard/service_stat
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DashServiceStatOutput'
              type: object
      summary: 服务统计
      tags:
      - 首页大盘
  /service/service_add_grpc:
    post:
      consumes:
      - application/json
      description: 添加GRPC服务
      operationId: /service/service_add_grpc
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddGRPCInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加GRPC服务
      tags:
      - 服务管理
  /service/service_add_http:
    post:
      consumes:
      - application/json
      description: 添加HTTP服务
      operationId: /service/service_add_http
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddHTTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加HTTP服务
      tags:
      - 服务管理
  /service/service_add_tcp:
    post:
      consumes:
      - application/json
      description: 添加TCP服务
      operationId: /service/service_add_tcp
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddTCPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加TCP服务
      tags:
      - 服务管理
  /service/service_delete:
    get:
      consumes:
      - application/json
      description: 服务删除
      operationId: /service/service_delete
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 服务删除
      tags:
      - 服务管理
  /service/service_detail:
    get:
      consumes:
      - application/json
      description: 服务详情
      operationId: /service/service_detail
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dao.ServiceDetail'
              type: object
      summary: 服务详情
      tags:
      - 服务管理
  /service/service_info:
    get:
      consumes:
      - application/json
      description: 服务信息列表
      operationId: /service/service_info
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 页数
        in: query
        name: page_no
        required: true
        type: integer
      - description: 每页个数
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceInfoOutput'
              type: object
      summary: 服务信息列表
      tags:
      - 服务管理
  /service/service_update_grpc:
    post:
      consumes:
      - application/json
      description: 修改GRPC服务
      operationId: /service/service_update_grpc
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateGRPCInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改GRPC服务
      tags:
      - 服务管理
  /service/service_update_http:
    post:
      consumes:
      - application/json
      description: 修改HTTP服务
      operationId: /service/service_update_http
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateHTTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改HTTP服务
      tags:
      - 服务管理
  /service/service_update_tcp:
    post:
      consumes:
      - application/json
      description: 修改TCP服务
      operationId: /service/service_update_tcp
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateTCPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改TCP服务
      tags:
      - 服务管理
swagger: "2.0"